requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-Bcrypt>=1.0.1
mysql-connector-python>=8.2.0
gunicorn>=22.0.0
